New item: "Luck Potion"! (See desc.)
Introduces a new item, the luck potion. This can be bought for 500 coins and will increase the chance of rare drops (diamonds and demon wings) by 10% for 10 minutes.

This Does NOT affect fishing, and only one can be active at a time. You may buy and own multiple of them at once.

You have to use the luck potion with the /use command for it to be active.